moved Slider from components to layout
