+++ src/layout/Slider/index.ts
